git-svn-id: https://iiop-net.svn.sourceforge.net/svnroot/iiop-net/trunk@1867 b113d865-ab28-0410-9ae9-d2690f63996d

+++ IIOPNet/IDLToCLSCompiler/IDLCompiler/IDLParserTokenManager.cs
@@ -1937,7 +1937,7 @@ public Token getNextToken()
      {
         if (jjmatchedPos + 1 < curPos)
            input_stream.backup(curPos - jjmatchedPos - 1);
-        if ((jjtoToken[jjmatchedKind >> 6] & ((ulong)(1L << (jjmatchedKind & 0x3F)))) != (ulong)0L)
+        if ((jjtoToken[jjmatchedKind >> 6] & ((ulong)((ulong)1L << (jjmatchedKind & 0x3F)))) != (ulong)0L)
         {
            matchedToken = jjFillToken();
        if (jjnewLexState[jjmatchedKind] != -1)
